refactor: make monitoring class
