+++ simpleor/__init__.py
@@ -1,20 +1,13 @@
 """Top-level package for simpleor."""
-import sys
-import logging
+import os
+from pathlib import Path
+
 import simpleor.base
 import simpleor.scheduler
 import simpleor.cli
 
 __author__ = """Lennart Damen"""
-__email__ = "lennart.damen@hotmail.com"
+__email__ = "lennart.damen.ai@gmail.com"
 __version__ = "0.0.4"
 
-
-logging.basicConfig(
-    stream=sys.stdout,
-    format="%(asctime)s | %(levelname)-8s | Process: %(process)d | %(name)s:%("
-    "funcName)s:%(lineno)d - %(message)s",
-    level=logging.DEBUG,
-)
-
-logger = logging.getLogger(f"{__name__}")
+PROJECT_DIRECTORY = Path(os.path.dirname(os.path.realpath(__file__))).parent
